requirements.txt
anyio==4.9.0
APScheduler==3.11.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dateparser==1.2.1
firebase-admin==6.5.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
joblib==1.5.1
msgpack==1.1.0
numpy==2.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
